requirements.txt
deta
fastapi
uvicorn
sqlalchemy
aiomysql
pymysql
mysql.connector
alembic
python-dotenv
datetime
pandas